Refactor iceberg/beam conversions for clarity
